Fixes segfault
